backend/pipelines/cadastral_parcels/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadastral_parcels_pipeline"
version = "0.1.0"
description = "Pipeline for processing Danish cadastral parcels data"
authors = [
    {name = "Landbruget.dk Team"}
]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.8.4",
    "backoff>=2.2.1",
    "duckdb>=0.9.0",
    "ibis-framework[duckdb]>=6.0.0",
    "pandas>=1.5.3",
    "geopandas>=0.13.0",
    "pyarrow>=12.0.0",
    "python-dotenv>=1.0.0",
    "shapely>=2.0.1",
    "tqdm>=4.65.0",
    "google-cloud-storage>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.0.215",
]

[tool.setuptools]
packages = ["bronze", "silver"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "B", "C4", "ARG", "SIM", "TID"]
ignore = ["E203", "E501", "W503"]